requirements.txt
asgiref==3.5.0
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.2
defusedxml==0.7.1
Django==3.2.12
django-cors-headers==3.13.0
django-debug-toolbar==3.3.0
django-debug-toolbar-template-profiler==2.0.2
django-extensions==3.1.5
django-filter==22.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
isort==5.10.1
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
psycopg2-binary==2.8.4
pycparser==2.21
pydotplus==2.0.2
pygraphviz==1.9
PyJWT==2.3.0
pyparsing==3.0.8
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.9
wrapt==1.14.1
